+++ packages/gamut/Markdown/overrides/Anchor/index.tsx
@@ -1,0 +1,33 @@
+/* eslint-disable jsx-a11y/anchor-has-content */
+import React, { FunctionComponent, HTMLAttributes } from 'react';
+import PropTypes from 'prop-types';
+
+export interface AnchorProps extends HTMLAttributes<HTMLAnchorElement> {
+  href: string;
+  title?: string;
+  height?: number;
+  width?: number;
+  target?: string;
+  rel?: string;
+}
+
+const Anchor: FunctionComponent<AnchorProps> = props => {
+  const anchorProps = {
+    ...props,
+    target: '_blank',
+    rel: 'noopener noreferrer',
+  };
+
+  if (typeof window !== 'undefined' && window.URL) {
+    const url = new window.URL(props.href, window.location.href);
+    // remove noopener/noreferrer on same origin urls
+    if (url.origin === window.location.origin) {
+      anchorProps.rel = '';
+      anchorProps.target = '';
+    }
+  }
+
+  return <a {...anchorProps} />;
+};
+
+export default Anchor;